Merged Modules and Lessons screens into one screen
This screen takes advantage of the Backdrop defined by the material design guidelines. In the background there is the list of modules and in the foreground there is the list of classes from the selected module.
Some changes were made to the classic guidelines. Among others, the most important is that on choosing a module, the backdrop foreground emerges automatically, hiding the background.